fix(apple): Prevent repeated IPC calls when not connected
Set an arbitrary limit of a single use when signed out
for the following operations that require IPC:

- log folder size calculation
- log export
- clearing log directory

The size will be cached unless user connected or cleared the logs.
The export and clearing buttons change to confirmation texts to subtly
show the user that it cannot be done again.

All actions are available again after the user reconnects successfully.