fix(portal): `phx-ignore` checkbox changes (#10879)
On Resources and Policies forms, we were triggering the form's
validation helpers when checking and unchecking a checkbox.
Unfortunately this causes the checkbox to be reset since it was not
saved across the to_form(changeset) rebuilding.

To prevent this, we simply ignore checkbox changes from triggering form
validations. We can also remove the `field=` property on these because
we are setting the `checked` property ourselves.

This will be refactored to be made simpler with the new modals approach,
so a minimal fix is implemented for now.

Related:
https://firezonehq.slack.com/archives/C098RV5BL1K/p1763024374383399
Fixes: #9143