fix(portal): use `[]:53` for ipv6 addresses
connlib expects the following format for IP addresses:

- ipv4:port
- ipv6
- [ipv6]:port

When storing the clients_upstream_dns values from the form in the admin
portal, we do not add brackets to IPv6 addresses, so URI.parse thinks
the address contains a port and strips the trailing `:` off.

We then sent these invalid IPs down to connlib which causes a
deserialization error.

To fix, we simply ensure IPs sent to connlib adhere to one of the
formats above.

Related: #10851