ci(rust): improve proptest coverage checks (#10918)
The current coverage checks for paths that we hit during our proptests
fail as soon as one of them is not satisfied. When iterating on the
proptests, it is useful to see in one go, which paths are currently not
hit to generate the missing regression seeds.

Hence, we refactor the script to perform all checks and fail if any of
them are not hit, outputting all missing ones.